reorganize cpp code, update:README
